requirements.txt

Flask==2.3.3
openai-whisper
moviepy
nltk
fpdf
